Change struct update to satisfy dialzyer
In its static analysis it was somehow losing track of the type of the
struct in `add_processed_tag/2` and thinking it was a expressing a
plain map.  A plain map breaks the contract.

The actual contract was not broken.  Introducing an `IO.inspect` in the
code showed a proper struct in the output... as well as resolving the
issue for dialyzer! None of this makes any sense unless there are
points where the struct information is lazily replaced and dialyzer does
not hit them.

Also not making a lot of sense is why it happens in the added call in
particular and not throughout the code.  Perhaps the multiple mutations
in a single function creates a too long indirection.

In any case, this is a pretty unsatisfying solution.  The anonymous
update function is probably more verbose than the maintainers would
like.  This commit is mostly published as a way to resolve the errors
and stimulate a discussion.