Port :erlang.bsr/2 to JS
This is something of an experiment for me.

This was largely (99+%) completed by AI tooling. It was given the
contents of the porting guidelines and produced the code herein.
I believe it has done a passable job after some coaxing.

I am happy to make additional changes to bring it more inline with
conventions and standards which the AI and/or I have misinterpreted.

I can also expand test coverage if there are specific edges to consider.

Closes #407